Update (#12)
* check and update terms cited in English

* check trifle

* update history of work sessions

* check trifle

* update presentation of all categories

Added quotes and capital for all categories: checked against English "reference" version.

* add spaces in terms written as one word

The English version is not consistent and this file is a reading version in which technical terms are already flagged by quotes and capitals. Same decision as in the online Spanish version.

* correct typo

* correct name of Processus object

Cf. the objects file.

* update readme about French translation

* integrate proofreading remarks

2017-12-01 remarks by Adeline Joffres (Huma-Num)

* update steps and roles

* Update readme.md

* Update readme.md

* Small corrections

I corrected one spelling from UK to US, as the latter seemed to be used throughout and eliminated one redundant word.

* fix typo

Co-authored-by: Christof Schöch <c.schoech@gmail.com>
Co-authored-by: Thibault Clérice <leponteineptique@gmail.com>
Co-authored-by: jody perkins <perkintj@miamioh.edu>
Co-authored-by: Brian Croxall <briancroxall@users.noreply.github.com>